refactor(notifications): split mobile app notification scheduler
Move Android scheduling, iOS scheduling, ring alarm handling, native adhan
playback, pre-notifications, and shared plugin setup into focused src/ files.

Preserve the public ScheduleAdhan API and keep ringAlarm as a top-level
@pragma('vm:entry-point') callback.

No behavior changes intended.
